+++ apps/bot/src/domains/guild/interactions/confirm-set-prefix.ts
@@ -3,19 +3,19 @@ import type { ButtonInteraction } from 'discord.js';
 import { ValidationError } from '../../../discord/errors.js';
 import type { ButtonHandler } from '../../../discord/types.js';
 import {
-  assertHasAdministratorPermission,
   assertInteractorIsTheOwner,
+  assertPermissionsHaveAdministrator,
   buildOpEmbed,
   parseOwnerDiscordId,
 } from '../../../discord/utils/index.js';
-import { CANCEL_SET_PREFIX_BUTTON_NAME, CONFIRM_SET_PREFIX_BUTTON_NAME } from '../constants.js';
+import { CONFIRM_SET_PREFIX_BUTTON_NAME } from '../constants.js';
 import { requireGuildId } from '../guards.js';
 import * as guildRepository from '../repository.js';
 
-async function confirmSetPrefix(interaction: ButtonInteraction, args: Array<string>): Promise<void> {
+async function handle(interaction: ButtonInteraction, args: Array<string>): Promise<void> {
   const ownerDiscordId = parseOwnerDiscordId(args[0]);
   assertInteractorIsTheOwner(interaction, ownerDiscordId);
-  assertHasAdministratorPermission(interaction.memberPermissions);
+  assertPermissionsHaveAdministrator(interaction.memberPermissions);
 
   const prefix = requireGuildPrefix(args[1]);
   const guildId = requireGuildId(interaction.guildId);
@@ -29,24 +29,6 @@ async function confirmSetPrefix(interaction: ButtonInteraction, args: Array<stri
   await interaction.editReply({ embeds: [embed], components: [] });
 }
 
-async function cancelSetPrefix(interaction: ButtonInteraction, args: Array<string>): Promise<void> {
-  const ownerDiscordId = parseOwnerDiscordId(args[0]);
-  assertInteractorIsTheOwner(interaction, ownerDiscordId);
-
-  await interaction.deferUpdate();
-  await interaction.editReply({ content: 'Annulé.', embeds: [], components: [] });
-}
-
-export const confirmSetPrefixButtonHandler: ButtonHandler = {
-  name: CONFIRM_SET_PREFIX_BUTTON_NAME,
-  handle: confirmSetPrefix,
-};
-
-export const cancelSetPrefixButtonHandler: ButtonHandler = {
-  name: CANCEL_SET_PREFIX_BUTTON_NAME,
-  handle: cancelSetPrefix,
-};
-
 function requireGuildPrefix(prefix: string | undefined): string {
   if (!prefix) {
     throw new ValidationError('Préfixe introuvable.');
@@ -54,3 +36,8 @@ function requireGuildPrefix(prefix: string | undefined): string {
 
   return prefix;
 }
+
+export const confirmSetPrefixButtonHandler: ButtonHandler = {
+  name: CONFIRM_SET_PREFIX_BUTTON_NAME,
+  handle,
+};
