+++ apps/bot/src/domains/guild/interactions/index.ts
@@ -1,5 +1,6 @@
 import type { ButtonHandler } from '../../../discord/types.js';
 
-import { cancelSetPrefixButtonHandler, confirmSetPrefixButtonHandler } from './set-prefix.js';
+import { cancelSetPrefixButtonHandler } from './cancel-set-prefix.js';
+import { confirmSetPrefixButtonHandler } from './confirm-set-prefix.js';
 
 export const guildButtonHandlers: Array<ButtonHandler> = [confirmSetPrefixButtonHandler, cancelSetPrefixButtonHandler];
