+++ packages/db/src/domains/guild/constants.ts
@@ -1,1 +1,2 @@
+export const MIN_GUILD_PREFIX_LENGTH = 1;
 export const MAX_GUILD_PREFIX_LENGTH = 8;
