+++ packages/db/src/domains/guild/index.ts
@@ -1,2 +1,2 @@
-export { MAX_GUILD_PREFIX_LENGTH } from './constants.js';
+export { MAX_GUILD_PREFIX_LENGTH, MIN_GUILD_PREFIX_LENGTH } from './constants.js';
 export { guild, type Guild } from './schema.js';
